fix: handler reference to use __dirname
