+++ src/handlers/event.user.created.ts
@@ -1,6 +1,9 @@
-module.exports.command = 'event.user.created';
+module.exports.command = "event.user.created";
 
 module.exports.listen = function(command, cb) {
-    const { bus } = this;
-    console.log(`User Created: ${command}`);
+  const { bus } = this;
+  console.log(`User Created: ${command}`);
+  cb();
 };
+
+console.log("test test");
